Added Rack::Cors, set to totally permissive
In general is not possible to access to an api server from a different
domain from the one where the api are running on. As we have 2 different domains for backend and frontend we need to set the HTTP Access Control (CORS) to be
permissive on our domains. For now, only for testing, I set CORS to be
totally permissive, this means that anyone from any host can access and
use our api server

Signed-off-by: Gabriele Proietti Mattia <gabry.gabry@hotmail.it>